# api.yaml
openapi: 3.0.3
info:
  title: Hospital Management System API
  description: >
    REST API for a hospital management system supporting three roles —
    Admin, Doctor, and Patient — with JWT-based authentication, appointment
    scheduling, treatment records, AI-powered patient summaries, and a
    post-appointment chat feature.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Unauthorized

    # ─── Doctor ──────────────────────────────────────────────────────────────
    Doctor:
      type: object
      properties:
        doctor_id:
          type: integer
        doctor_username:
          type: string
        doctor_name:
          type: string
        doctor_email:
          type: string
          format: email
        doctor_type:
          type: string
        doctor_qualification:
          type: string
        doctor_experience:
          type: string
        doctor_description:
          type: string
        dep_id:
          type: integer
          nullable: true
    DoctorCreateRequest:
      type: object
      required:
        - username
        - name
        - password
        - doctor_email
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
        type:
          type: string
        dep_id:
          type: integer
          nullable: true
        doctor_email:
          type: string
          format: email
        qualification:
          type: string
        experience:
          type: string
        description:
          type: string
    DoctorUpdateRequest:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
        type:
          type: string
        dep_id:
          type: integer
          nullable: true
        email:
          type: string
          format: email
        qualification:
          type: string
        experience:
          type: string
        description:
          type: string

    # ─── Patient ─────────────────────────────────────────────────────────────
    Patient:
      type: object
      properties:
        patient_id:
          type: integer
        patient_name:
          type: string
        patient_username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        gender:
          type: string
        age:
          type: integer
        address:
          type: string
        is_blacklisted:
          type: boolean
...
    Department:
      type: object
      properties:
        dep_id:
          type: integer
        dep_name:
          type: string
        dep_des:
          type: string
        no_docs_registered:
          type: integer
    DepartmentRequest:
      type: object
      required:
        - dep_name
      properties:
        dep_name:
          type: string
        dep_des:
          type: string

    # ─── Appointment ─────────────────────────────────────────────────────────
    Appointment:
      type: object
      properties:
        app_id:
          type: integer
        patient_id:
          type: integer
        doctor_id:
          type: integer
        date:
          type: string
          format: date
        time:
          type: string
          format: time
          example: "09:00"
        status:
          type: string
          enum: [BOOKED, COMPLETED, CANCELLED]
        visit_type:
          type: string
          enum: [Morning, Afternoon, Evening]
        token_no:
          type: integer
          nullable: true
          description: Queue position within the slot (patient dashboard only)

    # ─── Treatment ───────────────────────────────────────────────────────────
...
paths:
...
  /api/admin/dashboard:
    get:
      tags: [Admin]
      summary: Admin dashboard — all doctors, patients, and 10 upcoming appointments
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          doctors:
                            type: array
                            items:
                              $ref: '#/components/schemas/Doctor'
                          patients:
                            type: array
                            items:
                              $ref: '#/components/schemas/Patient'
                          recent_appointments:
                            type: array
                            items:
                              $ref: '#/components/schemas/Appointment'
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/doctors:
    get:
      tags: [Admin]
      summary: List all doctors (cached)
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          doctors:
                            type: array
                            items:
                              $ref: '#/components/schemas/Doctor'
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Admin]
      summary: Create a new doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorCreateRequest'
      responses:
        "201":
          description: Doctor created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          doctor:
                            $ref: '#/components/schemas/Doctor'
        "400":
          description: Username or email already exists, or department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/doctors/search:
    get:
      tags: [Admin]
      summary: Search doctors by name
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Partial name to search (case-insensitive)
      responses:
        "200":
          description: Matching doctors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          doctors:
                            type: array
                            items:
                              $ref: '#/components/schemas/Doctor'
  /api/admin/doctors/{doctor_id}:
    parameters:
      - in: path
        name: doctor_id
        required: true
        schema:
          type: integer
    get:
      tags: [Admin]
      summary: Get a single doctor's details
      responses:
        "200":
          description: Doctor details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          doctor:
                            $ref: '#/components/schemas/Doctor'
        "404":
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Admin]
      summary: Update a doctor's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorUpdateRequest'
      responses:
        "200":
          description: Doctor updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        "404":
          description: Doctor or new department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Admin]
      summary: Delete a doctor
      responses:
        "200":
          description: Doctor deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        "404":
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/patients/{patient_id}:
    parameters:
      - in: path
        name: patient_id
        required: true
        schema:
          type: integer
    get:
      tags: [Admin]
      summary: Get a patient's details and all their appointments
      responses:
        "200":
          description: Patient details with appointment history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          patient:
                            $ref: '#/components/schemas/Patient'
                          appointments:
                            type: array
                            items:
                              $ref: '#/components/schemas/Appointment'
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Admin]
      summary: Hard-delete a patient record
      responses:
        "200":
          description: Patient deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/admin/patients/{patient_id}/history:
    get:
      tags: [Admin]
      summary: Full appointment history for a patient (descending)
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          patient:
                            $ref: '#/components/schemas/Patient'
                          appointments:
                            type: array
                            items:
                              $ref: '#/components/schemas/Appointment'
  /api/admin/patients/{patient_id}/blacklist:
    post:
      tags: [Admin]
      summary: Blacklist a patient
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient blacklisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/admin/patients/{patient_id}/unblacklist:
    post:
      tags: [Admin]
      summary: Remove blacklist from a patient
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient unblacklisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/admin/departments:
    get:
      tags: [Admin]
      summary: List all departments (cached)
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          departments:
                            type: array
                            items:
                              $ref: '#/components/schemas/Department'
    post:
      tags: [Admin]
      summary: Create a new department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentRequest'
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        "400":
          description: Department name is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/departments/{dep_id}:
    parameters:
      - in: path
        name: dep_id
        required: true
        schema:
          type: integer
    get:
      tags: [Admin]
      summary: Get a department's details (cached)
      responses:
        "200":
          description: Department details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    put:
      tags: [Admin]
      summary: Update a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentRequest'
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    delete:
      tags: [Admin]
      summary: Delete a department
      responses:
        "200":
          description: Department deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/admin/appointments:
    get:
      tags: [Admin]
      summary: List all appointments with optional filters
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [BOOKED, COMPLETED, CANCELLED]
        - in: query
          name: doctor_id
          schema:
            type: integer
        - in: query
          name: patient_id
          schema:
            type: integer
      responses:
        "200":
          description: Filtered appointment list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          appointments:
                            type: array
                            items:
                              $ref: '#/components/schemas/Appointment'
...